Problema 3/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
